@@ -7,15 +7,21 @@
     },
     "devDependencies": {
         "@popperjs/core": "^2.11.6",
+        "@tailwindcss/postcss": "^4.1.8",
         "@vitejs/plugin-vue": "^5.2.4",
         "autoprefixer": "^10.4.21",
-        "axios": "^1.6.4",
+        "axios": "^1.9.0",
         "bootstrap": "^5.2.3",
         "laravel-vite-plugin": "^1.0.0",
         "postcss": "^8.5.4",
         "sass": "^1.56.1",
-        "tailwindcss": "^4.1.8",
         "vite": "^6.3.5",
-        "vue": "^3.2.37"
+        "vue": "^3.5.16",
+        "vue-loader": "^17.4.2"
+    },
+    "dependencies": {
+        "@tailwindcss/vite": "^4.1.8",
+        "tailwindcss": "^4.1.8",
+        "vue-router": "^4.5.1"
     }
 }
